// packages/shared/src/types.ts
export interface CapsuleTierInfo {
  id: string;
  name: string;
  displayName: string;
  description?: string;
  price: number;
  totalSupply: number;
  remaining: number;
  commonRate: number;
  rareRate: number;
  epicRate: number;
  legendaryRate: number;
  mythicRate: number;
  ezziRate: number;
  cssClass: string;
  animation: string;
}